feat: add view module
